webApp/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
attrs==25.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
blockbuster==1.5.24
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
cloudpickle==3.1.1
colorama==0.4.6
cryptography==44.0.2
dataclasses-json==0.6.7
distro==1.9.0
fastapi==0.115.12
Flask==3.1.0
flask-cors==5.0.1
forbiddenfruit==0.1.4
frozenlist==1.5.0
google-api-core==2.24.2
google-api-python-client==2.166.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
googleapis-common-protos==1.69.2
greenlet==3.1.1
grpcio==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema_rs==0.29.1
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-community==2.0.7
langchain-openai==0.3.6
langchain-text-splitters==0.3.8
langgraph==0.2.74
langgraph-api==0.1.0
langgraph-checkpoint==2.0.24
langgraph-cli==0.2.1
langgraph-runtime-inmem==0.0.2
langgraph-sdk==0.1.61
langsmith==0.3.10
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.26.1
mdurl==0.1.2
msgpack==1.1.0
multidict==6.4.3
mypy-extensions==1.0.0
numpy==2.2.4
openai==1.64.0
orjson==3.10.15
ormsgpack==1.9.1
packaging==24.2
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.30.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.10.6
pydantic-settings==2.8.1
pydantic_core==2.27.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.3
python-dotenv==1.0.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
rsa==4.9
sniffio==1.3.1
soupsieve==2.6
SpeechRecognition==3.14.3
SQLAlchemy==2.0.40
sse-starlette==2.1.3
starlette==0.46.1
structlog==25.2.0
tenacity==9.0.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.3.0
uvicorn==0.34.0
watchfiles==1.0.5
Werkzeug==3.1.3
yarl==1.19.0
zstandard==0.23.0
